+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - REMOVED
 - IMPROVED
   - Tooltip timeouts handling
+  - Lightbox rework for better behavior
 - BUGFIXES
 
 # 2.0.0-beta.5 - 2022-01-04